refactor(build): split the command bridge's non-body file handling
generate_commands.py reached 602 lines and test_generate_commands.py 608, both
over the 500-line taste ceiling, which the count ratchet caught as a
regression against baseline 566.

generate_commands_resources.py now owns everything that travels beside a
command's SKILL.md mirror: resource suffix validation, resource discovery, the
references/ tree walk, and the shared copy primitive. It also carries the
progressive-disclosure layout rationale, which is where a reader looking at the
copy code will want it. The bridge keeps frontmatter merging, collision
detection, and orchestration.

The references tests move to test_generate_commands_references.py. No behavior
change: 32 tests pass before and after the split.

Refs #4016
